Replace index.d.ts files with types.ts
